+++ src/components/dashboardGridConfig.ts
@@ -1,0 +1,12 @@
+export const gridConfig = {
+  paddingLeft: "16px",
+  paddingRight: "16px",
+  paddingTop: "2px",
+  paddingBottom: "2px",
+  
+  display: 'grid',
+  gridTemplateColumns: '32px 1fr 32px',
+  alignItems: 'start',
+  gap: '0 8px',
+};
+
